requirements.txt
# AlphaMark Flash Loan Application - Python Dependencies
# Version-pinned for production stability

# Web3 and Ethereum
web3>=6.20.0,<7.0.0
eth-account>=0.9.0,<1.0.0
eth-abi>=4.0.0,<5.0.0

# Networking
requests>=2.31.0,<3.0.0

# Redis for inter-service communication
redis>=5.0.0,<6.0.0

# Logging
concurrent-log-handler>=0.9.24,<1.0.0

# Data processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0.0,<7.0.0
aiohttp>=3.9.0,<4.0.0
